表示内容の変更と細かいコメントのアップデート

+++ libraries/DayCounterLib/days_paint.h
@@ -22,7 +22,7 @@
 
 class DaysPaint {
  public:
-  static void PaintDaysToFrameMemory(int days, epd::Epd *display);
+  static void PaintDaysToFrameMemory(uint16_t days, epd::Epd *display);
 
   // If |rollover| is zero, rollover will be deleted.
   // Valid |rollover| range is zero to 20.
